+++ src/main/java/mx/uach/videoclub/dao/jdbc/helper/VideoDaoJdbcHelper.java
@@ -32,7 +32,7 @@ public class VideoDaoJdbcHelper {
      * lo pedimos
      */
     public final static Director makeDirector(ResultSet rs) throws SQLException{
-        Director obj = new Director(rs.getInt(Director.FIELDS[0]), rs.getString(Director.FIELDS[1]));
+        Director obj = new Director(rs.getLong(Director.FIELDS[0]), rs.getString(Director.FIELDS[1]));
         return obj;
     }
     
@@ -45,7 +45,7 @@ public class VideoDaoJdbcHelper {
      * lo pedimos
      */
     public final static Actor makeActor(ResultSet rs) throws SQLException{
-        Actor obj = new Actor(rs.getString(Actor.FIELDS[1]), rs.getString(Actor.FIELDS[2]), rs.getInt(Actor.FIELDS[0]));
+        Actor obj = new Actor(rs.getString(Actor.FIELDS[1]), rs.getString(Actor.FIELDS[2]), rs.getLong(Actor.FIELDS[0]));
         return obj;
     }
     
@@ -60,8 +60,8 @@ public class VideoDaoJdbcHelper {
     public final static Cinta makeCinta(ResultSet rs) throws SQLException{
         Pelicula pelicula;
         VideoDao dao = new VideoDaoJDBC();
-        pelicula = (Pelicula) dao.getPeliculaById(rs.getInt(Cinta.FIELDS[2]));
-        Cinta obj = new Cinta(rs.getInt(Cinta.FIELDS[1]), pelicula, rs.getInt(Cinta.FIELDS[0]));
+        pelicula = (Pelicula) dao.getPeliculaById(rs.getLong(Cinta.FIELDS[2]));
+        Cinta obj = new Cinta(rs.getInt(Cinta.FIELDS[1]), pelicula, rs.getLong(Cinta.FIELDS[0]));
         return obj;
     }
     
@@ -76,8 +76,8 @@ public class VideoDaoJdbcHelper {
     public final static Ficha makeFicha(ResultSet rs) throws SQLException{
         Socio socio;
         VideoDao dao = new VideoDaoJDBC();
-        socio = (Socio) dao.getSocioById(rs.getInt(Ficha.FIELDS[2]));
-        Ficha obj = new Ficha(rs.getDate(Ficha.FIELDS[1]), socio, rs.getInt(Ficha.FIELDS[0]));
+        socio = (Socio) dao.getSocioById(rs.getLong(Ficha.FIELDS[2]));
+        Ficha obj = new Ficha(rs.getDate(Ficha.FIELDS[1]), socio, rs.getLong(Ficha.FIELDS[0]));
         return obj;
     }
     
@@ -91,9 +91,9 @@ public class VideoDaoJdbcHelper {
      */
     public final static Lista makeLista(ResultSet rs) throws SQLException{
         VideoDao dao = new VideoDaoJDBC();
-        Socio socio = dao.getSocioById(rs.getInt(Lista.FIELDS[2]));
-        Pelicula pelicula = dao.getPeliculaById(rs.getInt(Lista.FIELDS[1]));
-        Lista obj = new Lista(rs.getDate(Lista.FIELDS[3]), rs.getDate(Lista.FIELDS[4]), rs.getBoolean(Lista.FIELDS[5]), socio, pelicula, rs.getInt(Lista.FIELDS[0]));
+        Socio socio = dao.getSocioById(rs.getLong(Lista.FIELDS[2]));
+        Pelicula pelicula = dao.getPeliculaById(rs.getLong(Lista.FIELDS[1]));
+        Lista obj = new Lista(rs.getDate(Lista.FIELDS[3]), rs.getDate(Lista.FIELDS[4]), rs.getBoolean(Lista.FIELDS[5]), socio, pelicula, rs.getLong(Lista.FIELDS[0]));
         return obj;
     }
     
@@ -107,8 +107,8 @@ public class VideoDaoJdbcHelper {
      */
     public final static Pelicula makePelicula(ResultSet rs) throws SQLException{
         VideoDao dao = new VideoDaoJDBC();
-        Director director = dao.getDirectorById(rs.getInt(Pelicula.FIELDS[4]));
-        Pelicula obj = new Pelicula(rs.getString(Pelicula.FIELDS[1]), EGenero.valueOf(rs.getString(Pelicula.FIELDS[2])), rs.getInt(Pelicula.FIELDS[3]), director, rs.getInt(Pelicula.FIELDS[0]));
+        Director director = dao.getDirectorById(rs.getLong(Pelicula.FIELDS[4]));
+        Pelicula obj = new Pelicula(rs.getString(Pelicula.FIELDS[1]), EGenero.valueOf(rs.getString(Pelicula.FIELDS[2])), rs.getInt(Pelicula.FIELDS[3]), director, rs.getLong(Pelicula.FIELDS[0]));
         return obj;
     }
     
@@ -122,9 +122,9 @@ public class VideoDaoJdbcHelper {
      */
     public final static Prestamo makePrestamo(ResultSet rs) throws SQLException{
         VideoDao dao = new VideoDaoJDBC();
-        Ficha ficha = dao.getFichaById(rs.getInt(Prestamo.FIELDS[2]));
-        Cinta cinta = dao.getCintaById(rs.getInt(Prestamo.FIELDS[1]));
-        Prestamo obj = new Prestamo(rs.getDate(Prestamo.FIELDS[3]), EEstatusPrestamo.valueOf(rs.getString(Prestamo.FIELDS[4])), ficha, cinta, rs.getInt(Prestamo.FIELDS[0]));
+        Ficha ficha = dao.getFichaById(rs.getLong(Prestamo.FIELDS[2]));
+        Cinta cinta = dao.getCintaById(rs.getLong(Prestamo.FIELDS[1]));
+        Prestamo obj = new Prestamo(rs.getDate(Prestamo.FIELDS[3]), EEstatusPrestamo.valueOf(rs.getString(Prestamo.FIELDS[4])), ficha, cinta, rs.getLong(Prestamo.FIELDS[0]));
         return obj;
     }
     
@@ -137,7 +137,7 @@ public class VideoDaoJdbcHelper {
      * lo pedimos
      */
     public final static Socio makeSocio(ResultSet rs) throws SQLException{
-        Socio obj = new Socio(rs.getString(Socio.FIELDS[1]), rs.getString(Socio.FIELDS[2]), rs.getString(Socio.FIELDS[3]), rs.getInt(Socio.FIELDS[0]));
+        Socio obj = new Socio(rs.getString(Socio.FIELDS[1]), rs.getString(Socio.FIELDS[2]), rs.getString(Socio.FIELDS[3]), rs.getLong(Socio.FIELDS[0]));
         return obj;
     }
